+++ pages/moderator/AnalyticOrder.js
@@ -1,0 +1,14 @@
+import { Component } from "react";
+import { Text, View } from "react-native";
+
+class AnalyticOrder extends Component {
+
+    render() {
+
+        return (
+            <Text>AnalyticOrder</Text>
+        )
+    }
+}
+
+export default AnalyticOrder