+++ pages/moderator/CompletedOrderMod.js
@@ -1,0 +1,14 @@
+import { Component } from "react";
+import { Text, View } from "react-native";
+
+class CompletedOrderMod extends Component {
+
+    render() {
+
+        return (
+            <Text>CompletedOrderMod</Text>
+        )
+    }
+}
+
+export default CompletedOrderMod