+++ pages/moderator/TabsMod.js
@@ -9,6 +9,7 @@ import OrderMod from "./OrderMod";
 import ProfileMod from "./ProfileMod";
 import PromotionMod from "./PromotionMod";
 
+
 //Screen Names
 const catalogName = "Catalog"
 const orderName = "Order"